refinamentos
Refinamentos importantes para uma experiência melhor

+++ CRUD/lista-cargos.php
@@ -30,7 +30,7 @@ include_once './include/header.php';
               <td>'.$row['TetoSalarial'].'</td>
               <td>
                <a href="salvar-cargos.php?id='.$row['CargoID'].'" class="btn btn-edit">Editar</a>
-                <a href="./action/cargos.php?id='.$row['CargoID'].'&acao=excluir" class="btn btn-delete">Excluir</a>
+                <a href="./action/cargos.php?id='.$row['CargoID'].'&acao=excluir" class="btn btn-delete" onclick="return confirm(\'Tem certeza que deseja excluir?\')">Excluir</a>
                 </form>
                </td>
               </tr>';
