+++ CRUD/lista-producao.php
@@ -39,10 +39,10 @@ include_once './include/header.php';
             <td>'.$row['ProducaoID'].'</td>
             <td>'.$row['produto_nome'].'</td>
             <td>'.$row['funcionario_nome'].'</td>
-            <td>'.$row['DataProducao'].'</td>
+            <td>'.str_replace('-', '/', $row['DataProducao']).'</td>
             <td>
             <a href="salvar-producao.php?id='.$row['ProducaoID'].'" class="btn btn-edit">Editar</a>
-            <a href="./action/producao.php?id='.$row['ProducaoID'].'&acao=excluir" class="btn btn-delete">Excluir</a>
+            <a href="./action/producao.php?id='.$row['ProducaoID'].'&acao=excluir" class="btn btn-delete" onclick="return confirm(\'Tem certeza que deseja excluir?\')">Excluir</a>
             </td>
             </tr>
             ';
